management/requirements.txt
django-celery-beat==2.8.0
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.2
cffi==1.17.1
channels==4.2.2
psycopg2==2.9.10
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.4
djangorestframework==3.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.3
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.0.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
pillow==11.2.1
djangorestframework_simplejwt==5.5.0
openai==1.79.0
stripe==12.2.0
cloudinary==1.44.0
django-cloudinary-storage==0.3.0
django-redis==5.4.0